requirements.txt
speech_recognition==3.14.2
pyttsx3==2.90
requests==2.31.0
google-generative-ai==0.3.2